Move menus, uploading logic to new files
- uploading mutation is now stored in editor itself
- format imports
